refactor : state/effect/intent 파일 이동
